fix(core): un-break win64 clippy — RawSocket is already u64, the cast is same-type
CI's Windows clippy (-D warnings) rejects `raw as u64` in the qWAVE flow
guard: std's RawSocket is u64 on Windows, so the cast is a no-op
(clippy::unnecessary_cast). Verified with the CI's exact invocation
(cargo clippy -p punktfunk-host --features nvenc,amf-qsv -- -D warnings)
on the RTX box.

+++ crates/punktfunk-core/src/transport/qos_windows.rs
@@ -102,8 +102,9 @@ pub(super) fn add_media_flow(socket: &UdpSocket, class: MediaClass) -> Option<Qo
         return None;
     }
     // Construct the guard FIRST so an early return below still removes the flow membership.
+    // (`raw` is already `u64` — std's `RawSocket` — so no cast; win64 clippy denies same-type casts.)
     let flow = QosFlow {
-        socket: raw as u64,
+        socket: raw,
         flow_id,
     };
     // Pin the exact code point. Succeeds for elevated processes or under the "allow non-admin
